fix(reader): adjust page turn animation threshold and reset logic
Updated the threshold for turning pages in CoverPageAnimation to one third of the screen width. This change aims to improve the user experience by making it more responsive to user swipes.

Additionally, modified the _resetPosition method in _ReaderViewState to use reverse animation for resetting the page position. Removed duplicated call to MediaQuery, replacing it with a single variable holding the screen size.